ENH: Add build rules for sphinx documentation
Move files for sphinx documentation to Utilities/Sphinx, for consistency
with Utilities/Doxygen. Add build rules to invoke sphinx-build. Remove
unused reST file. Tweak module loading so that building the sphinx doc
does not depend on VTK's Python bindings (as (a) we don't have a way to
express this dependency, (b) setting up the paths to be able to import
the modules would be awkward, and (c) we shouldn't be trying to document
VTK anyway).