ENH: New version of Slicer logo
Create a new/updated version of the Slicer logo for use in the About
dialog. This one has a transparent background and white glow around the
text to ensure visibility against dark backgrounds. It also features
entirely reconstructed graphics for the main logo that is geometrically
correct (see also slicer-logo.py which was used to generate the
reference image) and has pixel-aligned elements (which produces much
crisper lines when rasterized).