ENH: New Slicer logo, take 2
Recreate the slicer logo (again) in a smaller size that more closely
matches the original default size of the About dialog. (The old dialog
was using a downsampled version of the original image!) This one has the
major benefit of also being an appropriate size for the splash screen,
which will enable us to recreate that as well.

To facilitate that, this version separates the logo from the text, and
combines the components via a Gimp .xcf file.

The (new) new version is essentially the same as the previous version,
tweaked for pixel alignment in the new size. The text has also been
hand-kerned for aesthetics and pixel alignment.