ENH: More tweaks to new logos
Adjust the XY and YZ plane grid elements to have less Y extent. This
more closely matches the proportion of the old logo (and also looks
better; already these are significantly elongated in the Y dimension
compared to true squares, so lessening that is an improvement), at the
trade-off of more 'empty' space along the Y axis versus the old logo.
(However the old logo also severely distorted the border between outer
and inner rings along the Y axis, which made the inner 'core' smaller
along the Y axis to begin with.)