COMP: Add VTKv6 Python path to launcher
It appears that the way VTKv6 chooses the location to which to install
its Python bindings has changed. As obtaining the Python version early
enough to override the path with the new VTKv6 mechanism appears to be
non-obvious/non-trivial, we instead punt, and add the location that is
now being used to the launcher's PYTHONPATH's.

Also, remove specification of VTK_INSTALL_PYTHON_USING_CMAKE, which no
longer exists in VTKv6.