COMP: Rewrite build rules for Python scripts
Create new slicerFunctionCompilePythonScript to replace use of
ctkMacroCompilePythonScript for generating rules to copy and compile
Python scripts. The visible effects are similar, with the major
differences that only named files are compiled and installed, and
compile rules are per file (rather than one giant 'compile everything we
can find'). Additionally, rules (and targets) are injected directly at
the call site, making ctkFunctionAddCompilePythonScriptTargets
unnecessary and removing the need to create additional targets in case
of generated input files.

Most crucially, the INSTALL_DESTINATION no longer has the last component
of DESTINATION appended, and copies individual files, not directories
(which may have cruft in them). By default, the directory structure of
the input names is maintained; additional paths to strip may be given
with the RELATIVE option.