ENH: Another iteration of the Slicer logo
Rescale the 'updated' Slicer logo so that it can be downscaled more
readily. The previous version looked good at its native 256x256
resolution; this version should look good at 64x64 and 128x128 as well
(the 'wire' elements still need manual kerning for each resolution, but
the rest downscales reasonably). As an unexpected bonus, the 'solid'
part of the logo also happens to be exactly half the overall width,
which means the elements are also directly usable for the icon version
of the logo as well.