ENH: New Slicer icon
Create SVG and new individual image files for the Slicer desktop icon,
based on the new version of the Slicer logo. The SVG is essentially the
same, with some components removed, and a few stroke widths tweaked.

Note that the 16x16 ("Small") version is not a direct export, but was
made by externally downscaling (in gimp, using cubic interpolation) a
512x512 version with all strokes removed. (Direct export at this size
was producing a result that was "muddier" and had some unexpected
transparency.)

The .ico file is not updated in this commit.