ENH: Update Windows .ico file; improve icons
Update the Windows .ico file with the new version of the icon. Tweak the
other icon images slightly from their raw-from-inkscape versions with
some judicious application of gimp's "unsharp mask" filter in order to
improve perceived sharpness, vibrancy, and contrast.

The 128x128 version (.ico only for now) comes from SlicerLogo.svg and
has the full grid and stroke widths from the same. The 64x64 version is
now a (raster) blend of the two, such that it has the full grid, but the
outlines are more subtle.