COMP: Create a functional install target
Update Artichoke to commontk/Artichoke@b0cf6fc85d5a. Use new
ExternalProject_Install_CMake to install external projects when doing a
superbuild build of slicer, so that there is a functional install target
for the same. Update CTK to bring in the recently added similar feature
from the same.

At least on Fedora, this is sufficient to produce a Slicer that can be
successfully run after building the 'install' target.