COMP: Automatically generate qSlicerBaseQTGUI.qrc
Add build rules to run genqrc.py at build time in order to ensure that
qSlicerBaseQTGUI.qrc is always up to date with respect to the currently
available icons. Embed icon file hashes in the generated .qrc to ensure
that the resource .cxx is regenerated when any of the input resource
files change, while removing explicit dependencies on the files from the
build system (which will break the build when a file is deleted).